backend/requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.5
django-crispy-forms==1.9.0
django-extensions==2.2.9
django-cors-headers==3.1.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
idna==2.9
money==1.3.0
pytz==2019.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.9
